Stage 1: Working GORM connection to the database
This commit contains the most basic HTTP server with no endpoints that
can connect to a MariaDB database, the credentials of which are
specified inside a .env file.

In the future, a README will be created with the instructions on how to
create a MariaDB database, how to manage Go dependencies and how to run
modify and run the project.